Reducing thesis abstract length to match ProQuest
According to an email I received about my submission, the abstract should be shortened to 350 words.
This was not written in any rules or guidelines I found, and this was explicitly contradicted in documentation I found on ProQuest's website (https://about.proquest.com/globalassets/proquest/files/pdf-files/preparing-your-manuscript.pdf).
This may be a new policy or something that hasn't changed globally at ProQuest, yet.

Due to the lack of email responses and my deadline for submission, I'm shortening the abstract to re-upload the document to ProQuest.

The original version is kept in the `_expandedabstract` PDF.